Move OptimizeMixedDistinctAggregations
Move class before refactoring to preserve history
